Add migrate file
